Convert tabs to spaces
Cloud9 uses 4 spaces by default.
If students copy and paste from a file with tabs into the editor, their file will have a
mix of tabs and spaces.
Python will complain if they are mixed on the same line, which can
happen as you edit and move stuff around.
So let's make it easy for them.