Renames user commands, clarifies wording
Changes the user commands from "print" and "print all" to "view item'" and "view list".

A few motivations for this: a natural extension for students to write is a mechanism to print the whole cart. Since that's more than printing a list, it's odd for the current "print all" not to print the whole cart. So renaming to "print item" and "print list" would make space for "print all" or "print cart".

However, think "view" is a more natural word for users to have in their command, since it is agnostic to the way the view happens (command-line, phone app, etc.) the user command "print" is better for putting things onto paper. Making the user commands "view *" also makes more clear the distinction between the user command "view item", "view list" and the student-defined functions print_list(), print_item().

For a reference implementation of the lab using these new terms, see https://repl.it/@davidrhmiller/TEALS2018-Lab-404-Shopping-List