Fix long lines in readings.md
One of the primary principles of Markdown is that the document is as
readable in its raw format as it is when rendered. To this end, lines
should be kept to a maximum length of 100 characters.

In addition, this fixes multiple instances where paragraphs were not
separated by blank lines, which collapses to a single paragraph in
Markdown.