Fix lesson 1.05 code example
We had problems linking to `code.py` since it failed to use an
explicitly-relative path to the Python file. On review, there's no good
reason to require a jump, so I added the four-line example to the bottom
of the lesson page and updated the link.

Fixes #3 - Lesson 1.05 code links are broken